daml/Dovetail-DAML-Client/activity/tocontract/activity.go
package parsecontract

import (
	"github.com/project-flogo/core/activity"
	logger "github.com/project-flogo/core/support/log"
)

var activityMd = activity.ToMetadata(&Settings{}, &Input{}, &Output{})
var log = logger.ChildLogger(logger.RootLogger(), "daml.activity.parsecontract")

func init() {
	_ = activity.Register(&Activity{}, New)
}

// Activity is a stub for creating a contract
type Activity struct {
}

// New creates a new Activity
func New(ctx activity.InitContext) (activity.Activity, error) {
	return &Activity{}, nil
}

// Metadata implements activity.Activity.Metadata
func (a *Activity) Metadata() *activity.Metadata {
	return activityMd
}

// Eval implements activity.Activity.Eval
func (a *Activity) Eval(ctx activity.Context) (done bool, err error) {
	input := &Input{}
	err = ctx.GetInputObject(input)
	if err != nil {
		return false, err
	}

	log.Debugf("input body = %v\n", input.Input)

	output := make([]interface{}, 0)
	switch t := input.Input.(type) {
	case []interface{}:
		output = append(output, t...)
	default:
		output = append(output, t)
	}

	result := Output{Output: output}
	ctx.SetOutputObject(&result)

	return true, nil
}
